+++ lib/theme.dart
@@ -6,6 +6,7 @@ import 'package:flutter/material.dart'
         ButtonThemeData,
         CardTheme,
         Color,
+        ColorScheme,
         Colors,
         DialogTheme,
         FloatingActionButtonThemeData,
@@ -103,12 +104,12 @@ const blue = Color(0xFF00B0FF);
 ThemeData getTheme(bool darkMode) {
   var color = darkMode ? grey : white;
 
-  // final ColorScheme colorSchema;
-  // if (darkMode) {
-  //   colorSchema = const ColorScheme.dark();
-  // } else {
-  //   colorSchema = const ColorScheme.light();
-  // }
+  final ColorScheme colorSchema;
+  if (darkMode) {
+    colorSchema = const ColorScheme.dark();
+  } else {
+    colorSchema = const ColorScheme.light();
+  }
   return ThemeData(
     fontFamily: 'Jost',
     brightness: darkMode ? Brightness.dark : Brightness.light,
@@ -134,9 +135,11 @@ ThemeData getTheme(bool darkMode) {
         darkMode ? color.c6.withOpacity(0.3) : color.c6.withOpacity(0.2),
     splashColor: color.c6.withOpacity(0.3),
     primaryColor: Colors.white,
-    // colorScheme: colorSchema.copyWith(
-    //   secondary: Colors.blue,
-    // ),
+    useMaterial3: true,
+    colorScheme: colorSchema.copyWith(
+      primary: Colors.blue,
+      secondary: Colors.blue,
+    ),
     buttonTheme: ButtonThemeData(
       minWidth: 85,
       height: 35,