+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 
 ## Next
+- Support adaptive icon
 - Auto start following system theme when switching to the system theme
-- Support adaptive icons
+- Use Material 3 theme
 
 ## 1.0.0 - 2019 Oct 19
 - Initial release